generalize vis_z_stack using get_slice -> show_stack
